Use ObjectSpace to locate models by their file names.
This approach eliminates necessity to guess model name by file name that
allow to process models with any kind of non-standard capitalization in
class and module names. Though, it still requires that underscored file
name matches underscored model class name. If file is located not at the
top level of models directory then we assume that model is either nested
in the corresponding chain of modules or is not nested at all.