Prevent root_dir option from being empty
When setting root_dir to an empty string (as is done in the default rake
task config) this line previously caused root_dir to become an empty
array.

This in turn caused factories, specs etc not to be annotated.

This is just a quick band-aid on a larger problem. We have option
parsing spread out over many different places, with slight mismatches
like this in the assumptions made.