Use a less error-prone way of specifying gem files (#662)
Before, whenever a new file is added to `lib`, the contributor or
maintainer has to remember to explicitly add it to the gemspec. If this
step is forgotten, it results in a broken release, like #657.

This commit makes this process less error prone by using `git ls-files`
to automatically include all appropriate files in the gem. This is in
line with what is advocated by Bundler's gem template:

https://github.com/bundler/bundler/blob/2a4d88a7044528209321fd4441e995ec66f604b9/lib/bundler/templates/newgem/newgem.gemspec.tt#L26