Add columns managed by Globalize gem (#602)
* Add columns managed by Globalize gem

Globalize hooks into the model and removes the translated columns
from the `klass.columns`. This commit checks if globalize is
hooked into the model and adds the necessary columns to the
annotation array.

* Disable Rubocop Metrics/BlockLength for spec files

RSpec spec files can contain long blocks easily because
of the outher describe methods. So this rule makes not too much
sense for these files.