Fixes updating columns when using alternative formats
Updates `column_pattern` in `AnnotateModels.annotate_one_file` to
support RDoc and Yard doc formats. `column_pattern` is responsible for
searching for existing annotations in the file - if the patten is not
matched in file, annotations are not updated.

Since the previous regular expression did not match the Yard
(`@!`) or RDoc (`**<attributed>**::`), when attempting to update
annotations for a model, it would look like there were no changes since
parsing the annotation blocks with the regular expression would return
nothing.