Fix updating when using Yard format
The lgoic in `AnnotateModels.annotate_one_file` for determining whether
there were changes a class' list of columns relied on a `column_pattern`
regular expression to fetch the list of columns in the annotation.

This regular expression was hard-coded to match on the original, default
"bare" syntax. If using an alternative documentation format (e.g.,
Yard), this regular expression would fail to find the columns and not
attempt to update the class' annoations.

This adds an additional regular expression to search for Yard-formatted
documentation.