Fix conflicting annotations for multiple-database scenario
Given a Rails application with multiple databases, tables in two
databases with the same table name, and corresponding models with
different names, some files get annotated with the wrong table's schema.

The issue is that some of the patterns include `%TABLE_NAME%`, which
will be identical between the two tables, even though they have
different model names.

This fixes the issue by eliminating the use of the table name when the
model is not connected to the primary database.

Fixes #891.