Fix retrieve_indexes_from_table when indexes is empty and base table does not exist. (#849)
Some tables may have a table_name_prefix but no indexes. Previous versions of
the code would strip the prefix and look for indexes on the resulting table 
which likely would not exist. This causes DB errors, at least in MySQL. So now
check if the new table exists first before trying to show its indexes.